monday

+++ app/Http/Controllers/Api/ResetPasswordController.php
@@ -3,8 +3,9 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\ResetPasswordRequest;
-use App\Http\Requests\ResetTokenRequest;
+use App\Http\Requests\Api\ResetPasswordRequest;
+use App\Http\Requests\Api\ResetTokenRequest;
+use App\Http\Requests\Api\UpdatePasswordRequest;
 use App\Models\ResetPassword;
 use App\Models\ResetToken;
 use App\Models\User;
@@ -33,17 +34,40 @@ class ResetPasswordController extends Controller
 
         return $this->successResponse(['code' => $resetCode]);
     }
+
     public function resettoken(ResetTokenRequest $request)
     {
-
         $user = User::where('email', $request->email)->first();
 
 
+        $user->reset_password_code()->delete();
+        $user->reset_password_token()->delete();
+
+        $token = Str::random(length: 20);
+
+        $user->reset_password_token()->create([
+            'email' => $request->email,
+            'token' => $token,
+        ]);
+
+        return $this->successResponse(['token' => $token]);
+    }
+    public function updatePassword(UpdatePasswordRequest $request)
+    {
+        $user = User::whereHas('reset_password_token', function ($query) use ($request) {
+            $query->where('token', $request->token);
+        })->first();
+        if(!$user){
+
+            return $this->errorResponse('error');
+
+        }
+        $user->password = $request->password;
+        $user->save();
 
         $user->reset_password_token()->delete();
-        $data['token'] = $user->createToken(request()->userAgent())->plainTextToken;
+        $user->reset_password_code()->delete();
 
-
-        return $this->successResponse($data);
+        return $this->successResponse('Password updated successfully.');
     }
 }
